+++ src/OpenLoco/src/GameCommands/Town/CreateTown.cpp
@@ -1,3 +1,4 @@
+#include "CreateTown.h"
 #include "Audio/Audio.h"
 #include "Date.h"
 #include "Economy/Expenditures.h"
