+++ src/OpenLoco/src/GameCommands/Town/RemoveTown.cpp
@@ -1,3 +1,4 @@
+#include "RemoveTown.h"
 #include "Audio/Audio.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/Buildings/RemoveBuilding.h"
