+++ src/OpenLoco/src/GameCommands/Town/RenameTown.cpp
@@ -1,3 +1,4 @@
+#include "RenameTown.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
 #include "Graphics/Gfx.h"
